keep function name and move also sim_get_default_cmake_builder to builder
